fix: resolve clippy warnings in mount_detect and stats
Remove unnecessary u32 cast and collapse nested if statements.

+++ crates/mt-tauri/src/cache/mount_detect.rs
@@ -42,7 +42,7 @@ mod macos {
                 return false;
             }
             // If MNT_LOCAL is NOT set, the filesystem is remote
-            (stat.f_flags as u32 & MNT_LOCAL) == 0
+            (stat.f_flags & MNT_LOCAL) == 0
         }
     }
 }
